Improve perimeter spreading and the GUI.
Allow choosing the filter type (rectangle, perimeter).
Fix the PSF visualizer (move the PSF correctly).
Add an option to linearly tone-map an LDR image in the picture box.
In the spreading filter allow forcing a limit on the PSF radius to be used (even though there might be pre-generated PSFs with greater radius).